fix errors when restoring debug dumps
the collection creation API changed in 3.12 and now complains about
invalid attributes that are specified upon collection creation.

for example, when restoring 3.11 debug dumps into 3.12, there can be
error messages logged as follows:

ERROR [ee638] {deprecation} The createCollection request contains an illegal combination and will be rejected in the future: {"errorNum":10,"errorMessage":"Found unexpected attribute 'objectId'"}

this PR removes a few invalid attributes from the debug dump, so that
they will not be used when creating the collection from the dump.