Fix a potential data race when running multiple read-only queries on the same streaming transaction. (#21077)
This race could implicitly occur when we try to start the query on the fast path in EngineInfoContainerDBServerServerBased, in case of a timeout send the cancel request and then start the new query on the slow path. It could happen that the setup request on the fast path had already inserted the query, but is still holding the shared_ptr. So the cancel request would actually find the query and remove it from the registry, but it would not yet be destroyed. Only once the shared_ptr is dropped would the query be destroyed and the memory tracker updated. But this would now race with the subsequent query on the slow path.
For read-only operations we don't actually track anything in the RocksDBMethods, so we now disable tracker updates for all read-only queries, so that actual concurrent usage should be safe. In addition, we drop the query pointer earlier when inserting it into the query registry.