fix a potential race when killing queries (#21075)
canceling an async job that executes an AQL query via the
RestJobHandler's `cancel` operation could lead to a race between the
RestJobHandler calling `cancel` on the RestCursorHandler which owns the
query, and the RestCursorHandler object that currently executed the
query.
the cancel operation could lead to the query being shut down while it
was operating, which was not race-free. instead of trying to attempt a
full shutdown during a cancel operation, we are now only setting the
kill flag on the coordinator in a soft way, which avoids all potential
races, but may lead to a later shutdown of the query. this is
potentially a good tradeoff, as the RestJobHandler's cancel operation
should be used only very rarely in practice.