Add TSan suppression for exception_ptr
The reference counting of exception_ptr is part of the prebuilt glibc and is
therefore not instrumented by TSan, so TSan is not aware of the synchronization
that happens there. That means we have to ignore all races in destructors of
exceptions as a result of a __exception_ptr release that causes the refcnt to
drop to zero.