populate "filtered" statistics in constrained sort executor (#21122)
* populate "filtered" statistics in constraint sort executor

the constrained sort executor actually can filter out data, because it
maintains a min- or max-heap and compares every input value with the
largest or smallest value on the heap. in case a value does not make it
into the heap, it is effectively filtered out. thus we should report all
values that are discarded like this as "filtered" in the statistics.

* fix tests

* revert invalid changes

---------

Co-authored-by: Lars Maier <lars@arangodb.com>