Fix potential sleeping barber in async prefetching (#20791)
* Fix potential sleeping barber in async prefetching

If an async prefetch task returns WAITING (e.g. from a RemoteNode), the wakeup could arrive at a point were the RestHandler has already been removed and is thus swallowed. When the next RestHandler makes the next call, we simply return the result from the prefetch task. But if that is WAITING, we now wait for a wakeup that will never come.
To fix that, the next call will check if the prefetch task has returned WAITING, and in that case will immediately make the next call to the upstream dependency. If the wakeup has already occurred, we will get the new data, otherwise we will get another WAITING, but then we know that the wakeup is still going to happen.

* Make sure each ExecutionEngine has its own SharedQueryState

* Small refactoring + add comments

* Allow async prefetching for nodes with upstream RemoteNode dependencies

* Adjust some tests

* Remove incorrect assertion

* Update CHANGELOG

* Fix schmutz

* Fix test: SharedQueryState needs to be set up properly

* Fix stuff

* Remove incorrect assertion

* Fix engine initialization

* Clean up includes