GraphManager support to find VertexCollections
Introduces a method to the graph manager that allows finding which
vertex collections are attached to a given edge collection by
inspecting all graphs in the database to find whether any of them has
the given edge collection in it.

This code assumes that these definitions are consistent among graphs
if an edge collection is shared between multiple named graphs.

No tests yet.