Fix BTS-2233
The problem was that the coroutine could resume in a separate thread
while the RestHandler would go into waiting state and give up its lock.

To fix that, we have separated the query registration and query
execution such that the query is only executed when the register
operation was not suspended, otherwise the query execution is resumed
via the usual wake up mechanism, which ensures proper locking.